+++ docs/history.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Values for environment `env-vars` now support context formatting
 - Add `name` override for environments to allow for regular expression matching
+- Display help text the first time the `shell` command is executed
 - Add support for Almquist (`ash`) shells
 
 ***Fixed:***